Refactored out reused functions.
