+++ detectors/Cargo.toml
@@ -9,3 +9,4 @@ if_chain = "1.0.2"
 scout-audit-clippy-utils = { version = "=0.2.3" }
 utils = { path = "../utils" }
 itertools = { version = "0.12" }
+common = { path = "common" }
