+++ detectors/iterators-over-indexing/Cargo.toml
@@ -10,6 +10,7 @@ crate-type = ["cdylib"]
 dylint_linting = { workspace = true }
 if_chain = { workspace = true }
 scout-audit-clippy-utils = { workspace = true }
+utils = { workspace = true }
 
 [package.metadata.rust-analyzer]
 rustc_private = true
