+++ detectors/zero-address/Cargo.toml
@@ -9,8 +9,8 @@ crate-type = ["cdylib"]
 [dependencies]
 dylint_linting = { workspace = true }
 if_chain = { workspace = true }
-
 scout-audit-clippy-utils = { workspace = true }
+utils = { workspace = true }
 
 [package.metadata.rust-analyzer]
 rustc_private = true
